+++ src/main.ts
@@ -1,6 +1,5 @@
 #!/usr/bin/env node
 import { Command } from 'commander';
-import path from 'path';
 import { checkNodeVersion } from './checkNodeVersion';
 import { authCommand } from './commands/auth';
 import { cacheCommand } from './commands/cache';
@@ -21,35 +20,8 @@ import { runDbMigrations } from './migrate';
 import { generateRedteamCommand } from './redteam/commands/generate';
 import { initCommand as redteamInitCommand } from './redteam/commands/init';
 import { pluginsCommand as redteamPluginsCommand } from './redteam/commands/plugins';
-import { type UnifiedConfig } from './types';
 import { checkForUpdates } from './updates';
-import { maybeReadConfig } from './util/config/load';
-
-export async function loadDefaultConfig(configNames: string[]): Promise<{
-  defaultConfig: Partial<UnifiedConfig>;
-  defaultConfigPath: string | undefined;
-}> {
-  const pwd = process.cwd();
-  let defaultConfig: Partial<UnifiedConfig> = {};
-  let defaultConfigPath: string | undefined;
-
-  // NOTE: sorted by frequency of use
-  const extensions = ['yaml', 'yml', 'json', 'cjs', 'cts', 'js', 'mjs', 'mts', 'ts'];
-
-  for (const ext of extensions) {
-    for (const name of configNames) {
-      const configPath = path.join(pwd, `${name}.${ext}`);
-      const maybeConfig = await maybeReadConfig(configPath);
-      if (maybeConfig) {
-        defaultConfig = maybeConfig;
-        defaultConfigPath = configPath;
-        return { defaultConfig, defaultConfigPath };
-      }
-    }
-  }
-
-  return { defaultConfig, defaultConfigPath };
-}
+import { loadDefaultConfig } from './util/config/default';
 
 async function main() {
   await checkForUpdates();
